Reorganize server files
